+++ common/__init__.py
@@ -1,0 +1,2 @@
+from .Board import Board
+from .BoardStatus import BoardStatus
